Add Support for Weighted Grading
Switched return type back from int to double

+++ GradeBook/GradeBooks/BaseGradeBook.cs
@@ -107,7 +107,7 @@ namespace GradeBook.GradeBooks
             }
         }
 
-        public virtual int GetGPA(char letterGrade, StudentType studentType)
+        public virtual double GetGPA(char letterGrade, StudentType studentType)
         {
             var additionalPoints = 0;
